ENH: Save Adjacent Pixel Value as Cell Data
When a mesh is derived from a segmentation image, it is frequently useful
to know, for each cell in the mesh, the value of the adjacent pixel in
the segmentation image.  In this way, it is possible to extract a labeled
mesh from a multi-label segmentation.  This patch adds an option (off by
default) which allows the value of the adjacent pixel to be saved as cell
data.