BUG: Regenerate Hardcoded Array with CC Bugfix
There was a recent bugfix to the ITK connected components filter,
which is useed to generate the hardcoded connected components array.
Unfortunately, the array was generated with the filter prior to
applying the most recent ITK patch.  This filter regenerates the array.
Additionally, it updates to the correct number of points for the test
cases.  Finally, it enables itk::QuadEdgeMesh in the test suite,
verifying that the filter now works for this data type on all the
tested inputs.